// src/event/dto/event-filter.dto.ts
export interface EventFilterDto {
  status?: string
  filesystemId?: string
  trackingStartedAt?: string
  birthTime?: string
  fileType?: string
  versionNumber?: number
  osUserId?: string
  executablePath?: string
  operationType?: string
  firstAt?: string
  limit?: number
  page?: number
  process?: string
}